[BackupHelper] Do not throw exception on unhandled events
Parsing events that aren't handled aren't issues such as reaching a
comment or whitespace. Just like most other places that the system
parses xmls we can just ignore events that we don't have interest in.
AOSP's implementation today doesn't have an issue with this but this
could cause issues with other serializer/parser implementations if a
manufacturer or frameworks maintainers choose to change the
implementation such as binary xml format.

Test: Push backup file with newlines inserted and restore
Bug: 299690524
Change-Id: Ie982d1afcdf8856204889cd218d6be74250cc8b2