Throw exception when APK install fails
Problem:

Many of our CTS tests install test helper APKs. But, in many cases,
if the install fails, we don't throw an error saying so; the install
silently fails and the test continues on.

When this happens, the test usually fails later, but it won't be
obvious that a package installation failure was the root cause.

This fix:

Detect when APK installation fails, and throw an exception.

Bug: 300325645
Test: cd packages/modules/Permission && atest :alltests
Change-Id: I9dd5df2004efa27dfb15cb94472c334ded740ca4