Improve error messages for permission assertions
When we assert the results from requestPermissions and checkPermission (
which we do frequently in the permissions test suite) we currently have
cryptic stack traces because the line numbers never match (due to
inlining) and the assertions have no (or poor) messages attached to
them.

Improve the error message so assertion stack traces are less cryptic.

Bug: 299676001
Test: cd packages/modules/Permission && atest :alltests
Change-Id: I153ecb7745d5724d028fe34e8ed4dbf6ff5a9b29