Make permission grant flow device aware
Device-aware permissions can guard access to resources on remote devices. CAMERA and RECORD_AUDIO are the first two permissions that need to be device aware. This change makes the existing permission grant flow device aware by passing the target device Id to the grant activity and fetch permission states for the remote device in the data preparation.

Bug: 292252664
Test: In ag/25223675
Change-Id: I0644d30a884d187efddedfc86f2bdf402226ed8b