Don't create unnecessary exception stacktrace objects
Creating an exception with stacktrace is an expensive
operation and we should not create them each time
a source is added. During privacy page loading (on test
device) i can see thousands of these objects are created.
Putting the stack traces behind debug flag, live datas
are kind of mature now and imo it shold be fine putting
the traces behind a debug flag.

Bug: 301681210
Bug: 300269033
Test: presubmit
Change-Id: Ifc7ccba9728b1c6865f8e979b5f448e434693e5d