Fix Intent resolution ordering.
The intended order was to have the implicit intent checked first.
However, we changed it to check the explicit intent first in
ag/21033470 as resolveActivity always returned a value regardless
of whether the non-exported internal activity actually resolved.

By checking manually if the activity is exported, we can use
the original (intended) order, which fixes an issue with the
overriding of the OTA source intent by setting a higher priority.

Test: atest SafetyCenterFunctionalTestCases
Bug: 296053629
Bug: 282920658
Relnote: Fix intent resolution ordering for Safety Center, which
allows overridng OTA source intent by setting higher priority.

Change-Id: I332200401340deb48f9ebd23d841f4e01c9d9e0d