Use RequiresApi annotation directly.
The lint is failing on CL, this appears to be due to the
RequiresApiTiramisuByDefault annotation.

Looks like the android.annotation.RequiresApi annotation
can be used on a package directly.

I've checked this on a CL and it appears to be WAI.
Let's try to submit this CL and check if the lint errors do
go away. Otherwise I'll revert back to using RequiresApi
on every class.

Test: lint works
Bug: 303836855
Relnote: N/A
Change-Id: Ied51bd506ec8a8ea59bbf63edd06423939df9a74