Add "Edit Photos" button in settings
When clicked, this button will open the photo picker, and allow users to
change which photos the app in question has access to. This also
consolidates the photo picker launching logic.

Bug: 303834669
Test: atest PhotoPickerPermissionTest
Relnote: add a button that allows users to select photos in settings
Change-Id: I8ca75795644915428db8e1b7fda5f15672818720
Merged-In: I8ca75795644915428db8e1b7fda5f15672818720