bugfix: Match behavior with the OS for granting Health permissions
Disallow granting Health permissions to incompatible apps in UI
as well. This is already disallowed in Health Fitness module for apps
not satisfying the requirements for granting Health permissions.

Apps requesting Health permissions in Android 14 are required to provide
an activity showing the rationale for granting its permission, based on
the following documentations:

https://developer.android.com/health-and-fitness/guides/health-connect/develop/get-started#show-privacy-policy

https://developer.android.com/reference/android/health/connect/HealthPermissions