[Role Logic Move] Shim RoleControllerServiceImpl
This implements the ability to toggle between running RoleControllerServiceImpl (and therefore the majority of the Role Business Logic) in either System Server or PermissionController, depending on whether the "roleControllerInSystemServer" flag is set.

New classes/interfaces:

- RoleControllerLocal: Runs logic directly in System Server
- RoleControllerRemote: Calls PermissionController to run logic
- RoleController: Interface to allow us to toggle between the above two

Also, RoleControllerServiceImpl is moved to role-controller, and given a
shim constructor, so that it can be directly invoked from RoleService.

Bug: 302563971
Test: atest CtsRoleTestCases
Change-Id: Ibc41b97682bb1c4f0199384b3687322cda436df8