[Role Logic Move] Call SS, not PC
Currently, the methods:

- RoleManager::isRoleVisible
- RoleManager::isApplicationVisibleForRole

...directly call corresponding RoleControllerManager methods. These are
the only two methods in RoleManager that call RoleControllerManager.

But, RoleControllerManager calls the RoleControllerService API which
runs in PermissionController, and, we want to invoke this logic in
System Server.

So, call RoleService instead, and have RoleService conditionally route
to either System Server or PermissionController depending on the flag
state.

Bug: 309139048
Test: atest CtsRoleTestCases
Change-Id: I417637269d70e28bd2bebcabc85f9ac6a033b0f8