[Role Logic Move] Update proguard
RoleParser uses reflection to load certain classes (*RoleBehavior).
Because of this, proguard thinks they're unused, and strips them out. In
PermissionController, we have a proguard rule to prevent this, but
when we enable the Role Logic Move feature flag, this code will be
compiled into Permission/service, not PermissionController. So, copy the
proguard rule to the Permission/service module.

Test: manual
Fix: 311235380
Change-Id: I7c7e82085ee2c93e474fcc1d2775573321cab2e8